Moved handleGet to a separate file.
